chore: move main package to api package so preload can import it as well
for now it was importing main using ../

related to https://github.com/podman-desktop/podman-desktop/pull/15485
